Code refactor to make it more modular
